+++ code/Terrain/Editor/SharpFallOff.cpp
@@ -13,13 +13,13 @@ namespace traktor::terrain
 
 T_IMPLEMENT_RTTI_CLASS(L"traktor.terrain.SharpFallOff", SharpFallOff, IFallOff)
 
-double SharpFallOff::evaluate(double x, double y) const
+float SharpFallOff::evaluate(float x, float y) const
 {
-	const double d = x * x + y * y;
-	if (d >= 1.0)
-		return 0.0;
+	const float d = x * x + y * y;
+	if (d >= 1.0f)
+		return 0.0f;
 	else
-		return 1.0;
+		return 1.0f;
 }
 
 }
